Menu item service projection added

+++ menu-service/src/main/java/com/restaurant/menu_service/dto/menu/MenuItemDto.java
@@ -13,7 +13,7 @@ public class MenuItemDto {
     private String foodType;
     private BigDecimal basePrice;
     private Integer preparationTime;
-    private Integer calories;
+    private Long calories;
     private String image;
     private Boolean available;
     private Boolean featured;
